# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api SQ Systems",
        "description": "Get the API SQ Systems",
        "version": "TEST VERSION"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Get Bearer token",
                "operationId": "2e9c317eea06b322356cd346eb0e29fb",
                "requestBody": {
                    "description": "Login request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Entity"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registration new user",
                "operationId": "7d8c5938c0f8c4cee857ce8e9a12d028",
                "requestBody": {
                    "description": "Registration request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Entity"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get information of user",
                "description": "Current user information",
                "operationId": "40a3b039a569fac2ba27e847ad38d515",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deleting a user",
                "description": "Deleting a user",
                "operationId": "63c211a82841480575d423406f3554d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get collection of users",
                "description": "Get collection of users",
                "operationId": "3726d86f8e220eecebc640a2d98fd22f",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "a928bfa10780d8115bd97093521d55bd",
                "requestBody": {
                    "description": "Registration request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "LoginRequest",
                "description": "Authorization request",
                "properties": {
                    "email": {
                        "description": "User Email",
                        "type": "string",
                        "example": "user@test.com"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "title": "UserRequest",
                "description": "Registration request",
                "properties": {
                    "name": {
                        "description": "User Name",
                        "type": "string",
                        "example": "Ivan"
                    },
                    "email": {
                        "description": "User Email",
                        "type": "string",
                        "example": "user@test.com"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "AuthResource": {
                "title": "UserResource",
                "description": "Authorization",
                "properties": {
                    "token": {
                        "description": "Authorization token",
                        "type": "string",
                        "example": "38|MaXb2t7AYhK5w6r1GpSpJPRYQEHaNZP7vBSIPWjt98361465"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "UserResource",
                "description": "User",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "User Name",
                        "type": "string",
                        "example": "Ivan"
                    },
                    "email": {
                        "description": "User Email",
                        "type": "string",
                        "example": "user@test.com"
                    },
                    "created_at": {
                        "description": "Date of Creation",
                        "type": "string",
                        "example": "2025-05-25 17:14:30"
                    },
                    "updated_at": {
                        "description": "Date of Update",
                        "type": "string",
                        "example": "2025-05-25 17:14:30"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}